+++ examples/main.py
@@ -1,3 +1,4 @@
+import argparse
 import functools
 
 import persist_to_disk as ptd
@@ -27,7 +28,7 @@ class FakeModel:
         print(f"Training model with lr={lr} for {echo} epochs.")
 
 @ptd.persistf(groupby=['dataset', 'epochs'], expand_dict_kwargs=['model_kwargs'], skip_kwargs=['device'], switch_kwarg='cache')
-def train_a_model(dataset, model_cls, model_kwargs, lr, epochs=30, device='cpu',**kwargs):
+def train_a_model(dataset, model_cls, model_kwargs, lr, epochs=30, device='cpu', args:argparse.Namespace=None, **kwargs):
     # This function demos how to use persist_to_disk in training the model
     model = model_cls(**model_kwargs)
     model.to(device)
@@ -73,3 +74,7 @@ if __name__ == '__main__':
     train_a_model('MNIST', FakeModel, {}, lr=0.001, cache=ptd.RECACHE)
 
     test_func(1)
+
+    # In v0.0.7, we can pass argparse.Namespace to the function.
+    print("\n\nIn v0.0.7, we can pass argparse.Namespace to the function.")
+    train_a_model('MNIST', FakeModel, {}, lr=0.001, args=argparse.Namespace(other_argument=30))